Extract a new file
